make image rendering for function help optional
