+++ doc/plugins/eu.esdihumboldt.cst.doc.functions.dynamic/src/eu/esdihumboldt/cst/doc/functions/internal/context/FunctionContextProvider.java
@@ -26,7 +26,6 @@ import eu.esdihumboldt.cst.doc.functions.FunctionReferenceConstants;
 import eu.esdihumboldt.cst.doc.functions.internal.toc.FunctionTopic;
 import eu.esdihumboldt.hale.common.align.extension.function.FunctionDefinition;
 import eu.esdihumboldt.hale.common.align.extension.function.FunctionUtil;
-import eu.esdihumboldt.hale.common.instance.orient.ONameUtil;
 import eu.esdihumboldt.hale.doc.util.context.SingleTopicContext;
 
 /**
@@ -34,8 +33,8 @@ import eu.esdihumboldt.hale.doc.util.context.SingleTopicContext;
  * 
  * @author Simon Templer
  */
-public class FunctionContextProvider extends AbstractContextProvider implements
-		FunctionReferenceConstants {
+public class FunctionContextProvider extends AbstractContextProvider
+		implements FunctionReferenceConstants {
 
 	/**
 	 * Context function topic
